+++ tests/LDUserTest.php
@@ -9,96 +9,140 @@ class LDUserTest extends \PHPUnit_Framework_TestCase {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->build();
         $this->assertEquals("foo@bar.com", $user->getKey());
-        $json = $user->toJSON();
-        $this->assertEquals("foo@bar.com", $json['key']);
     }
 
     public function testCoerceLDUserKey() {
         $builder = new LDUserBuilder(3);
         $user = $builder->build();
         $this->assertEquals("string", gettype($user->getKey()));
-        $json = $user->toJSON();
-        $this->assertEquals("string", gettype($json['key']));
     }
 
     public function testEmptyCustom() {        
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->build();
-        $json = $user->toJSON();
-        $this->assertTrue(!isset($json['custom']));
     }
 
     public function testLDUserSecondary() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->secondary("secondary")->build();
         $this->assertEquals("secondary", $user->getSecondary());
-        $json = $user->toJSON();
-        $this->assertEquals("secondary", $json['secondary']);
+    }
+
+    public function testLDUserPrivateSecondary() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateSecondary("secondary")->build();
+        $this->assertEquals("secondary", $user->getSecondary());
+        $this->assertEquals(array("secondary"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserIP() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->ip("127.0.0.1")->build();
         $this->assertEquals("127.0.0.1", $user->getIP());
-        $json = $user->toJSON();
-        $this->assertEquals("127.0.0.1", $json['ip']);
+    }
+
+    public function testLDUserPrivateIP() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateIp("127.0.0.1")->build();
+        $this->assertEquals("127.0.0.1", $user->getIP());
+        $this->assertEquals(array("ip"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserCountry() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->country("US")->build();
         $this->assertEquals("US", $user->getCountry());
-        $json = $user->toJSON();
-        $this->assertEquals("US", $json['country']);
+    }
+
+    public function testLDUserPrivateCountry() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateCountry("US")->build();
+        $this->assertEquals("US", $user->getCountry());
+        $this->assertEquals(array("country"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserEmail() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->email("foo+test@bar.com")->build();
         $this->assertEquals("foo+test@bar.com", $user->getEmail());
-        $json = $user->toJSON();
-        $this->assertEquals("foo+test@bar.com", $json['email']);
+    }
+
+    public function testLDUserPrivateEmail() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateEmail("foo+test@bar.com")->build();
+        $this->assertEquals("foo+test@bar.com", $user->getEmail());
+        $this->assertEquals(array("email"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserName() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->name("Foo Bar")->build();
         $this->assertEquals("Foo Bar", $user->getName());
-        $json = $user->toJSON();
-        $this->assertEquals("Foo Bar", $json['name']);
+    }
+
+    public function testLDUserPrivateName() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateName("Foo Bar")->build();
+        $this->assertEquals("Foo Bar", $user->getName());
+        $this->assertEquals(array("name"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserAvatar() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->avatar("http://www.gravatar.com/avatar/1")->build();
         $this->assertEquals("http://www.gravatar.com/avatar/1", $user->getAvatar());
-        $json = $user->toJSON();
-        $this->assertEquals("http://www.gravatar.com/avatar/1", $json['avatar']);
+    }
+
+    public function testLDUserPrivateAvatar() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateAvatar("http://www.gravatar.com/avatar/1")->build();
+        $this->assertEquals("http://www.gravatar.com/avatar/1", $user->getAvatar());
+        $this->assertEquals(array("avatar"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserFirstName() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->firstName("Foo")->build();
         $this->assertEquals("Foo", $user->getFirstName());
-        $json = $user->toJSON();
-        $this->assertEquals("Foo", $json['firstName']);
+    }
+
+    public function testLDUserPrivateFirstName() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateFirstName("Foo")->build();
+        $this->assertEquals("Foo", $user->getFirstName());
+        $this->assertEquals(array("firstName"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserLastName() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->lastName("Bar")->build();
         $this->assertEquals("Bar", $user->getLastName());
-        $json = $user->toJSON();
-        $this->assertEquals("Bar", $json['lastName']);
+    }
+
+    public function testLDUserPrivateLastName() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateLastName("Bar")->build();
+        $this->assertEquals("Bar", $user->getLastName());
+        $this->assertEquals(array("lastName"), $user->getPrivateAttributeNames());
+    }
+
+    public function testLDUserCustom() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->customAttribute("foo", "bar")->customAttribute("baz", "boo")->build();
+        $this->assertEquals(array("foo" => "bar", "baz" => "boo"), $user->getCustom());
+    }
+
+    public function testLDUserPrivateCustom() {
+        $builder = new LDUserBuilder("foo@bar.com");
+        $user = $builder->privateCustomAttribute("foo", "bar")->privateCustomAttribute("baz", "boo")->build();
+        $this->assertEquals(array("foo" => "bar", "baz" => "boo"), $user->getCustom());
+        $this->assertEquals(array("foo", "baz"), $user->getPrivateAttributeNames());
     }
 
     public function testLDUserAnonymous() {
         $builder = new LDUserBuilder("foo@bar.com");
         $user = $builder->anonymous(true)->build();
         $this->assertEquals(true, $user->getAnonymous());
-        $json = $user->toJSON();
-        $this->assertEquals(true, $json['anonymous']);
-
     }
 
     public function testLDUserBlankKey() {
@@ -116,4 +160,3 @@ class LDUserTest extends \PHPUnit_Framework_TestCase {
         $this->assertFalse($user->isKeyBlank());
     }
 }
-